requirements.txt
fastapi
uvicorn
langgraph
langchain-groq
langchain-core
langchain-community
langchain-experimental
langchain-huggingface
langchain-mongodb
langgraph-checkpoint-mongodb
motor
pymongo
sentence-transformers
pydantic
python-dotenv
requests
mcp[sse]
wikipedia
duckduckgo-search
pypdf
nest_asyncio
typing-extensions
passlib
bcrypt
pyjwt
python-multipart
